requirements.txt
Flask==2.2.2
requests==2.28.1
python-telegram-bot==13.13
gunicorn==20.1.0